+++ tests/integ/modin/test_faster_pandas.py
@@ -266,6 +266,55 @@ def test_agg(session, func):
         assert_series_equal(snow_result4, native_result4, check_dtype=False)
 
 
+@pytest.mark.parametrize(
+    "func",
+    [
+        "first",
+        "last",
+        "rank",
+        "shift",
+        "cumcount",
+        "cumsum",
+        "cummin",
+        "cummax",
+        "any",
+        "all",
+        "unique",
+    ],
+)
+@sql_count_checker(query_count=3)
+def test_groupby_no_param_functions(session, func):
+    with session_parameter_override(
+        session, "dummy_row_pos_optimization_enabled", True
+    ):
+        # create tables
+        table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+        session.create_dataframe(
+            native_pd.DataFrame([[2, 12], [2, 11], [3, 13]], columns=["A", "B"])
+        ).write.save_as_table(table_name, table_type="temp")
+
+        # create snow dataframes
+        df = pd.read_snowflake(table_name).sort_values("B", ignore_index=True)
+        snow_result = getattr(df.groupby("A")["B"], func)()
+
+        # verify that the input dataframe has a populated relaxed query compiler
+        assert df._query_compiler._relaxed_query_compiler is not None
+        assert df._query_compiler._relaxed_query_compiler._dummy_row_pos_mode is True
+        # verify that the output dataframe also has a populated relaxed query compiler
+        assert snow_result._query_compiler._relaxed_query_compiler is not None
+        assert (
+            snow_result._query_compiler._relaxed_query_compiler._dummy_row_pos_mode
+            is True
+        )
+
+        # create pandas dataframes
+        native_df = df.to_pandas()
+        native_result = getattr(native_df.groupby("A")["B"], func)()
+
+        # compare results
+        assert_series_equal(snow_result, native_result, check_dtype=False)
+
+
 @sql_count_checker(query_count=5, union_count=1)
 def test_concat(session):
     with session_parameter_override(